requirements.txt
flask~=3.1.0
openpyxl~=3.1.2
python-dotenv~=1.0.1
markdown~=3.5.2
markupsafe~=2.1.3

# LangChain ecosystem
langchain~=0.1.16
langchain-openai~=0.1.6
langchain-community~=0.0.34

# OpenAI SDK for LangChain to work
openai~=1.30.1
